business/routing/invariants

+++ oscar-cbls/src/main/scala/oscar/cbls/business/routing/invariants/NodesOfVehicle.scala
@@ -33,10 +33,10 @@ object NodesOfVehicle{
    */
   def apply(routes:ChangingSeqValue,v:Int,includeVehicleNode : Boolean = true):Array[CBLSSetVar] = {
     val model = routes.model
-    val emptySet = SortedSet.empty[Long]
+    val emptySet = SortedSet.empty[Int]
     val domain = routes.domain
 
-    val nodesOfVehicle = Array.tabulate(v+1L)((vehicle:Int) =>
+    val nodesOfVehicle = Array.tabulate(v+1)((vehicle:Int) =>
       CBLSSetVar(model,
         emptySet,
         domain,
@@ -59,20 +59,20 @@ class NodesOfVehicle(routes:ChangingSeqValue,
   includeVehicleNode : Boolean = true)  //there is actually one more vehicle, for unrouted nodes.
   extends Invariant() with SeqNotificationTarget{
 
-  val v = nodesOfVehicleOrUnrouted.length-1L
-  val n = routes.maxValue+1L
+  val v = nodesOfVehicleOrUnrouted.length-1
+  val n = routes.maxValue+1
 
   registerStaticAndDynamicDependency(routes)
   finishInitialization()
   nodesOfVehicleOrUnrouted.foreach(_.setDefiningInvariant(this))
 
-  private val savedValues:Array[SortedSet[Long]] = null
+  private val savedValues:Array[SortedSet[Int]] = null
   private var savedCheckpoint:IntSequence = null
   //TODO: use magic array here
   private val movedNodesSinceCheckpointArray:Array[Boolean] = Array.fill(n)(false)
-  private var movedNodesSinceCheckpointList:QList[Long] = null
-  private val vehicleOfNodeAtCheckpointForMovedPoints:Array[Long] = Array.fill(n)(0L)
-  private val vehicleOfNodeAfterMoveForMovedPoints:Array[Long] = Array.fill(n)(0L)
+  private var movedNodesSinceCheckpointList:QList[Int] = null
+  private val vehicleOfNodeAtCheckpointForMovedPoints:Array[Int] = Array.fill(n)(0)
+  private val vehicleOfNodeAfterMoveForMovedPoints:Array[Int] = Array.fill(n)(0)
 
   affect(computeValueFromScratch(routes.value))
 
@@ -87,7 +87,7 @@ class NodesOfVehicle(routes:ChangingSeqValue,
     val newValue = changes.newValue
 
     changes match {
-      case SeqUpdateInsert(value : Long, pos : Int, prev : SeqUpdate) =>
+      case SeqUpdateInsert(value : Int, pos : Int, prev : SeqUpdate) =>
         //on which vehicle did we insert?
         if(!digestUpdates(prev)) return false
         val insertedVehicle = RoutingConventionMethods.searchVehicleReachingPosition(pos,newValue,v)
@@ -134,7 +134,7 @@ class NodesOfVehicle(routes:ChangingSeqValue,
       case SeqUpdateLastNotified(value:IntSequence) =>
         true //we are starting from the previous value
       case SeqUpdateDefineCheckpoint(prev,isStarMode,checkpointLevel) =>
-        if(checkpointLevel == 0L) {
+        if(checkpointLevel == 0) {
           if (!digestUpdates(prev)) {
             affect(computeValueFromScratch(changes.newValue))
           }
@@ -148,7 +148,7 @@ class NodesOfVehicle(routes:ChangingSeqValue,
 
         if(checkpoint == null) false //it has been dropped following a Set
         else {
-          if(checkpointLevel == 0L) {
+          if(checkpointLevel == 0) {
             require(checkpoint quickEquals savedCheckpoint)
             restoreCheckpoint()
             true
@@ -181,7 +181,7 @@ class NodesOfVehicle(routes:ChangingSeqValue,
     }
   }
 
-  private def recordMovedPoint(node:Long, oldVehicle:Long, newVehicle:Long){
+  private def recordMovedPoint(node:Int, oldVehicle:Int, newVehicle:Int){
     require(oldVehicle != newVehicle)
     if(savedCheckpoint!= null) {
       if (!movedNodesSinceCheckpointArray(node)) {
@@ -193,22 +193,22 @@ class NodesOfVehicle(routes:ChangingSeqValue,
     }
   }
 
-  private def affect(value:Array[SortedSet[Long]]){
-    var currentV = 0L
+  private def affect(value:Array[SortedSet[Int]]){
+    var currentV = 0
     while(currentV <= v){
       nodesOfVehicleOrUnrouted(currentV) := value(currentV)
-      currentV += 1L
+      currentV += 1
     }
   }
 
-  private def computeValueFromScratch(s:IntSequence):Array[SortedSet[Long]] = {
-    val toReturn = Array.fill(v+1L)(SortedSet.empty[Long])
-    toReturn(v) = toReturn(v) ++ (v to n-1L)
+  private def computeValueFromScratch(s:IntSequence):Array[SortedSet[Int]] = {
+    val toReturn = Array.fill(v+1)(SortedSet.empty[Int])
+    toReturn(v) = toReturn(v) ++ (v to n-1)
     val it = s.iterator
-    var currentVehicle:Long = it.next()
-    require(currentVehicle == 0L)
+    var currentVehicle:Int = it.next()
+    require(currentVehicle == 0)
     if (includeVehicleNode)
-      toReturn(0L) = toReturn(0L) + (0L)
+      toReturn(0) = toReturn(0) + (0)
 
     while(it.hasNext){
       val node = it.next()
@@ -226,13 +226,13 @@ class NodesOfVehicle(routes:ChangingSeqValue,
 
   override def checkInternals(c : Checker) : Unit = {
     val values = computeValueFromScratch(routes.value)
-    for (vehicle <- 0L to v){
+    for (vehicle <- 0 to v){
       require(nodesOfVehicleOrUnrouted(vehicle).value equals values(vehicle), Some("error on vehicle " + v + " output-incremental:" + (nodesOfVehicleOrUnrouted(vehicle).value.diff(values(vehicle))) + " correct-output:" + (values(vehicle).diff(nodesOfVehicleOrUnrouted(vehicle).value))))
     }
 
     if(savedCheckpoint != null) {
       val nodesOfVehicleFromScratch = computeValueFromScratch(savedCheckpoint)
-      for (node <- 0L until n) {
+      for (node <- 0 until n) {
         if(movedNodesSinceCheckpointArray(node))
           require(nodesOfVehicleFromScratch(vehicleOfNodeAtCheckpointForMovedPoints(node)).contains(node))
       }
